+++ changelog.md
@@ -2,8 +2,8 @@
 
 ### [Latest]
 - Fix handling of nan values in histograms [#125](https://github.com/umami-hep/puma/pull/125)
-
 - Adding support for under- and overflow bins in `puma.HistogramPlot` [#124](https://github.com/umami-hep/puma/pull/124)
+- (Documentation) Adding copy-button to code cells in documentation [#131](https://github.com/umami-hep/puma/pull/131)
 
 ### [v0.1.8] (2022/08/30)